% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{Launch Model Builder app}
\usage{
build_model(...)
}
\arguments{
\item{...}{Optional [PM_data] and/or [PM_model] object(s). *PM_data* objects
supply covariates. *PM_model* objects supply any other defined model element,
and covariates only if there is no *PM_data* object or it has no covariates.
If the *PM_model* object contains covariates, they will be superseded by those in
the *PM_data* object, if supplied.}
}
\value{
Launches the shiny app.
}
\description{
Open the shiny model builder app.
}
\details{
The app will open in a separate window.
}
\author{
Michael Neely
}
