% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPM_model.R
\name{plot.PM_model}
\alias{plot.PM_model}
\title{Plot PM_model objects}
\usage{
\method{plot}{PM_model}(x, marker = TRUE, line = TRUE, explicit, implicit, ...)
}
\arguments{
\item{x}{The name of an [PM_model] object.}

\item{marker}{Controls the characteristics of the compartments (nodes).
It can be boolean or a list.
`TRUE` will plot the compartments with default characteristics.
`FALSE` will suppress compartment plotting.
If a list, can control some marker characteristics, including overriding defaults.
These include:
\itemize{
\item{`color`} Marker color (default: dodgerblue).
\item{`opacity`} Ranging between 0 (fully transparent) to 1 (fully opaque). Default is 0.5.
\item{`size`} Relative size of boxes, ranging from 0 to 1.  Default is 0.25.
\item{`line`} A list of  additional attributes governing the outline for filled shapes, most commonly
color (default: black) and width (default: 0.5).
}
<br>
<br>
Example: `marker = list(color = "red", opacity = 0.8, line = list(color = "black", width = 1))`}

\item{line}{Controls characteristics of arrows (edges).
`TRUE` will plot default lines. `FALSE` will suppress lines.
If a list, can control some line characteristics, including overriding defaults.
These include:
\itemize{
\item{`color`} Line color (default: black)
\item{`width`} Thickness in points (default: 1).
}
<br>
<br>
Example: `line = list(color = "red", width = 2)`}

\item{explicit}{A data frame or tibble containing two columns named `from` and `to`
to add additional connecting arrows to the plot indicating transfer between
compartments. For each row, the `from` column contains the compartment number of the arrow origin, and the
`to` column contains the compartment number of the arrow destination. Use 0 to indicate
a destination to the external sink. e.g., `explicit = data.frame(from = 3, to = 0)`}

\item{implicit}{Similar to `explicit`, used to add dashed connecting arrows
to the plot indicating implicit transfer between
compartments. For each row, the `from` column contains the compartment number of the arrow origin, and the
`to` column contains the compartment number of the arrow destination. Use 0 to indicate
a destination to the external sink. e.g., `implicit = data.frame(from = 2, to = 4)`}

\item{...}{Not used.}
}
\value{
Plots the object.
}
\description{
`r lifecycle::badge("stable")`

Plots a [PM_model] based on differential equations using network plots from tidygraph and ggraph packages.
}
\details{
This accepts a [PM_model] object and creates a network plot where nodes are compartments
and edges are arrows connecting compartments.
}
\examples{
library(PmetricsData)
NPex$model$plot()
}
\seealso{
[PM_model], [ggraph::ggraph()], [ggplot2::ggplot()]
}
\author{
Markus Hovd, Julian Otalvaro, Michael Neely
}
\concept{PMplots}
