% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PmetricsApps-package.R
\docType{package}
\name{PmetricsApps-package}
\alias{PmetricsApps}
\alias{PmetricsApps-package}
\title{PmetricsApps: Companion Applications for Pmetrics Package}
\description{
This package contains Shiny apps that serve as user interfaces for the Pmetrics package as well as the plotting function for models.
}
\author{
\strong{Maintainer}: Michael Neely \email{mneely@usc.edu}

}
\keyword{internal}
